backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
razorpay==1.4.2
python-multipart==0.0.20
pydantic==2.8.2
pydantic-settings==2.6.1
slowapi==0.1.9
psycopg2-binary==2.9.9
gunicorn==23.0.0
